+++ examples/table/demos/tree.vue
@@ -2,7 +2,7 @@
   <div>
     <div>
       <t-button @click="appendToRoot">添加根节点</t-button>
-      <t-button theme="default" style="margin-left: 16px" @click="setData1">重置数据</t-button>
+      <t-button theme="default" style="margin-left: 16px" @click="resetData">重置/更新数据</t-button>
       <t-button theme="default" style="margin-left: 16px" @click="onRowToggle">任意节点展开/收起</t-button>
       <t-button theme="default" style="margin-left: 16px" @click="onExpandAllToggle">{{
         expandAll ? '收起全部' : '展开全部'
@@ -95,6 +95,7 @@ function getData(currentPage = 1) {
             ...obj,
             id: thirdIndex,
             key: `我是 ${thirdIndex}_${currentPage} 号`,
+            list: true,
           };
         });
         return secondObj;
@@ -223,8 +224,9 @@ export default {
 
   methods: {
     // 全新赋值
-    setData1() {
+    resetData() {
       this.data = getData();
+      this.$refs.table.resetData(this.data);
     },
 
     // 更新
@@ -285,7 +287,7 @@ export default {
           type: 'Number',
         },
       ];
-      this.$refs.table.appendTo(row.key, newData);
+      this.$refs.table.appendTo(row?.key, newData);
       MessagePlugin.success(`已插入子节点我是 ${randomKey1} 和 ${randomKey2} 号，请展开查看`);
     },
 
@@ -387,6 +389,9 @@ export default {
         needed: key % 4 === 0 ? '是' : '否',
         description: '数据源',
       });
+
+      // 同时添加多个元素，示例代码有效勿删
+      // this.appendMultipleDataTo();
     },
 
     onAbnormalDragSort(params) {