+++ src/date-picker/panel/PanelContent.tsx
@@ -22,6 +22,7 @@ export default defineComponent({
     month: Number,
     tableData: Array,
     time: String,
+    popupVisible: Boolean,
     firstDayOfWeek: Number,
     partial: String,
     onYearChange: Function,
@@ -49,8 +50,8 @@ export default defineComponent({
     const { COMPONENT_NAME, defaultTimeValue, timeFormat } = this;
 
     return (
-      <div class={`${COMPONENT_NAME}--content`}>
-        <div class={`${COMPONENT_NAME}--${this.mode}`}>
+      <div class={`${COMPONENT_NAME}-content`}>
+        <div class={`${COMPONENT_NAME}-${this.mode}`}>
           <TDateHeader
             {...{
               props: {
@@ -59,7 +60,7 @@ export default defineComponent({
                 month: this.month,
                 onMonthChange: (val: number) => this.onMonthChange?.(val, { partial: this.partial }),
                 onYearChange: (val: number) => this.onYearChange?.(val, { partial: this.partial }),
-                onJumperClick: (val: number) => this.onJumperClick?.(val, { partial: this.partial }),
+                onJumperClick: ({ trigger }: { trigger: string }) => this.onJumperClick?.({ trigger, partial: this.partial }),
               },
             }}
           />
@@ -80,14 +81,16 @@ export default defineComponent({
         </div>
 
         {this.enableTimePicker && (
-          <div class={`${COMPONENT_NAME}--time`}>
-            <div class={`${COMPONENT_NAME}--time-viewer`}>{this.time || defaultTimeValue}</div>
+          <div class={`${COMPONENT_NAME}-time`}>
+            <div class={`${COMPONENT_NAME}-time-viewer`}>{this.time || defaultTimeValue}</div>
             <TTimePickerPanel
               {...{
+                key: this.partial,
                 props: {
                   format: timeFormat,
                   value: this.time,
                   onChange: this.onTimePickerChange,
+                  isShowPanel: this.popupVisible,
                   ...this.timePickerProps,
                 },
               }}