+++ src/select/select.tsx
@@ -95,8 +95,8 @@ export default defineComponent({
     const hasSlotOptions = ref(false);
     const focusing = ref(false);
     const labelInValue = ref(valueType.value === 'object');
-    const realValue = ref(keys.value?.value || 'value');
-    const realLabel = ref(keys.value?.label || 'label');
+    const realValue = computed(() => keys.value?.value || 'value');
+    const realLabel = computed(() => keys.value?.label || 'label');
     const realOptions = ref([] as Array<TdOptionProps>);
     const hoverIndex = ref(-1);
     const popupOpenTime = ref(250);
@@ -201,7 +201,7 @@ export default defineComponent({
       if (filterable.value) {
         // 仅有filterable属性时，默认不区分大小写过滤label
         return realOptions.value.filter(
-          (option) => option[realLabel.value].toString().toLowerCase().indexOf(tInputValue.value?.toString().toLowerCase())
+          (option) => option[realLabel.value]?.toString().toLowerCase().indexOf(tInputValue.value?.toString().toLowerCase())
             !== -1,
         );
       }
@@ -261,6 +261,7 @@ export default defineComponent({
 
     const debounceOnRemote = debounce(() => {
       instance.emit('search', tInputValue.value);
+      props.onSearch?.(tInputValue.value.toString());
     }, 300);
 
     watch(
@@ -424,9 +425,20 @@ export default defineComponent({
           break;
       }
     };
+
+    // 为 eventListener 加单独的 sync watch，以防组件在卸载的时候未能正常清除监听 (https://github.com/Tencent/tdesign-vue/issues/1170)
+    watch(
+      visible,
+      (val) => {
+        val && document.addEventListener('keydown', keydownEvent);
+        !val && document.removeEventListener('keydown', keydownEvent);
+      },
+      {
+        flush: 'sync',
+      },
+    );
+    // 其余逻辑使用默认 pre watch
     watch(visible, (val) => {
-      val && document.addEventListener('keydown', keydownEvent);
-      !val && document.removeEventListener('keydown', keydownEvent);
       !val && (showCreateOption.value = false);
       !val && tInputValue.value && setTInputValue('');
     });
@@ -525,6 +537,7 @@ export default defineComponent({
       setTInputValue('');
       visible.value = false;
       instance.emit('clear', { e });
+      props.onClear?.({ e });
     };
     const getOptions = (option: OptionInstance) => {
       // create option值不push到options里
@@ -553,17 +566,21 @@ export default defineComponent({
     };
     const createOption = (value: string) => {
       instance.emit('create', value);
+      props.onCreate?.(value);
     };
     const focus = (value: string, context: { e: FocusEvent }) => {
       focusing.value = true;
       instance.emit('focus', { value, e: context?.e });
+      props.onFocus?.({ value, e: context?.e });
     };
     const blur = (value: string, context: { e: FocusEvent | KeyboardEvent }) => {
       focusing.value = false;
       instance.emit('blur', { value, e: context?.e });
+      props.onBlur?.({ value, e: context?.e });
     };
     const enter = (value: string, context: { e: KeyboardEvent }) => {
       instance.emit('enter', { value, e: context?.e, inputValue: tInputValue.value });
+      props.onEnter?.({ value, e: context?.e, inputValue: tInputValue.value.toString() });
     };
     const getOverlayElm = (): HTMLElement => {
       let r;
@@ -617,7 +634,6 @@ export default defineComponent({
     onUpdated(() => {
       initOptions();
     });
-
     provide('tSelect', {
       getOptions,
       reachMaxLimit,
@@ -693,6 +709,7 @@ export default defineComponent({
   render(h) {
     const {
       multiple,
+      keys,
       autoWidth,
       bordered,
       readonly,
@@ -700,7 +717,6 @@ export default defineComponent({
       clearable,
       tDisabled,
       borderless,
-      empty,
       showCreateOption,
       displayOptions,
       isGroupOption,
@@ -737,6 +753,8 @@ export default defineComponent({
     const valueDisplay = this.renderValueDisplay(h);
     const placeholderText = this.getPlaceholderText();
     const prefixIcon = () => renderTNode('prefixIcon');
+    const empty = renderTNode('empty');
+
     const collapsedItems = () => renderCollapsedItems();
     const { overlayClassName, ...restPopupProps } = popupProps || {};
     return (
@@ -749,6 +767,7 @@ export default defineComponent({
           readonly={readonly}
           allowInput={showFilter}
           multiple={multiple}
+          keys={keys}
           value={selectedValue}
           valueDisplay={valueDisplay}
           clearable={clearable}
@@ -801,6 +820,8 @@ export default defineComponent({
             realLabel={realLabel}
             realValue={realValue}
             scroll={scroll}
+            panelTopContent={this.panelTopContent}
+            panelBottomContent={this.panelBottomContent}
           />
         </SelectInput>
       </div>